# Perf improvements to Span and Memory