--- a/src/System.Private.CoreLib/shared/System/ReadOnlyMemory.cs
+++ b/src/System.Private.CoreLib/shared/System/ReadOnlyMemory.cs
@@ -77,8 +77,13 @@ namespace System
                 this = default;
                 return; // returns default
             }
+#if BIT64
+            if ((ulong)(uint)start + (ulong)(uint)length > (ulong)(uint)array.Length)
+                ThrowHelper.ThrowArgumentOutOfRangeException();
+#else
             if ((uint)start > (uint)array.Length || (uint)length > (uint)(array.Length - start))
                 ThrowHelper.ThrowArgumentOutOfRangeException();
+#endif
 
             _object = array;
             _index = start;
@@ -172,10 +177,13 @@ namespace System
             // Used to maintain the high-bit which indicates whether the Memory has been pre-pinned or not.
             int capturedLength = _length;
             int actualLength = _length & RemoveFlagsBitMask;
-            if ((uint)start > (uint)actualLength || (uint)length > (uint)(actualLength - start))
-            {
+#if BIT64
+            if ((ulong)(uint)start + (ulong)(uint)length > (ulong)(uint)actualLength)
                 ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.start);
-            }
+#else
+            if ((uint)start > (uint)actualLength || (uint)length > (uint)(actualLength - start))
+                ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.start);
+#endif
 
             // Set the high-bit to match the this._length high bit (1 for pre-pinned, 0 for unpinned).
             return new ReadOnlyMemory<T>(_object, _index + start, length | (capturedLength & ~RemoveFlagsBitMask));