--- a/src/System.Private.CoreLib/shared/System/ReadOnlyMemory.cs
+++ b/src/System.Private.CoreLib/shared/System/ReadOnlyMemory.cs
@@ -78,6 +78,7 @@ namespace System
                 return; // returns default
             }
 #if BIT64
+            // See comment in Span<T>.Slice for how this works.
             if ((ulong)(uint)start + (ulong)(uint)length > (ulong)(uint)array.Length)
                 ThrowHelper.ThrowArgumentOutOfRangeException();
 #else
@@ -178,6 +179,7 @@ namespace System
             int capturedLength = _length;
             int actualLength = _length & RemoveFlagsBitMask;
 #if BIT64
+            // See comment in Span<T>.Slice for how this works.
             if ((ulong)(uint)start + (ulong)(uint)length > (ulong)(uint)actualLength)
                 ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.start);
 #else