This improves the performance of `MemoryExtensions.AsSpan<T>`, `Span<T>..ctor`, and `Span<T>.Slice` (and the read-only and `Memory<T>` equivalents) when the caller is responsible for providing both a start index and an offset parameter. Currently this is implemented via two branches that the JIT can't always elide. With this PR, on 64-bit machines, this is implemented as a single branch _and_ allows the JIT to perform constant-time folding where applicable, which offers performance gains in span-heavy parsing / formatting code.

As an example, I've provided metrics for `Guid.Parse(string) : Guid` and `WebUtility.HtmlEncode(string) : string`. `Guid.Parse` uses `Span.Slice` heavily, which is shown by the substantial increase in overall throughput between the baseline and this PR. `WebUtility.HtmlEncode` uses `ValueStringBuilder` under the covers, which uses `Span.Slice` internally but is a bit more conservative with those calls, so you'll see some throughput increase but nothing quite as substantial as `Guid`'s code did.

|     Method |       Toolchain |                value |     Mean |     Error |    StdDev | Ratio |
|----------- |---------------- |--------------------- |---------:|----------:|----------:|------:|
| **Guid_Parse** |        **baseline (27012-03)** | **2153e(...)c7c13 [36]** | **23.62 ms** | **0.3638 ms** | **0.3225 ms** |  **1.00** |
| Guid_Parse | experimental | 2153e(...)c7c13 [36] | 13.61 ms | 0.1180 ms | 0.1104 ms |  0.58 |
|            |                 |                      |                      |          |           |           |       |
| **HtmlEncode** |        **baseline (27012-03)** | **&amp;&lt;&gt;&amp;&lt;(...)&lt;&gt;&amp;&lt;&gt; [48]** | **71.50 ms** | **0.9397 ms** | **0.8331 ms** |  **1.00** |
| HtmlEncode | experimental | &amp;&lt;&gt;&amp;&lt;(...)&lt;&gt;&amp;&lt;&gt; [48] | 68.85 ms | 1.0039 ms | 0.9391 ms |  0.96 |