@GrabYourPitchforks 2018-11-02
I opened a separate linked issue (https://github.com/dotnet/coreclr/issues/20776) to help the JIT optimize this even further. In my brief testing this JIT optimization would further improve `Guid.Parse` battery performance by approx. another 10%.