@benaadams 2018-11-03
Is it worth making this a helper method somewhere? e.g.
```csharp
[MethodImpl(MethodImplOptions.AggressiveInlining)]
internal static bool IsRangeInBounds(int offset, int length, int availableLength)
{
#if BIT64
    // Since offset and length are both 32-bit, their sum can be computed across a 64-bit domain
    // without loss of fidelity. The cast to uint before the cast to ulong ensures that the
    // extension from 32- to 64-bit is zero-extending rather than sign-extending. The end result
    // of this is that if either input is negative or if the input sum overflows past Int32.MaxValue,
    // that information is captured correctly in the comparison against the availableLength field.
    // We don't use this same mechanism in a 32-bit process due to the overhead of 64-bit arithmetic.
    return ((ulong)(uint)offset + (ulong)(uint)length > (ulong)(uint)availableLength) ? false : true;
#else
    return ((uint)offset > (uint)availableLength || (uint)length > (uint)(availableLength - offset)) ? false : true;
#endif
}
```