@ahsonkhan 2018-11-03
> Is it worth making this a helper method somewhere?

Yes, that would definitely be helpful, but it may not be feasible as @GrabYourPitchforks stated.

> and I couldn't figure out how to move it into a separate helper method without getting the JIT to emit the typical "setb al / movzx eax, al" anti-pattern.

Maybe writing it the way @benaadams suggested (i.e. `return condition ? false : true` rather than just `return condition`) could help work around the JIT issue. Could be worth a shot seeing what code the JIT emits in that case. @AndyAyersMS, if that exercise fails, do you have some other suggestions that could work while we wait for a resolution for issues like https://github.com/dotnet/coreclr/issues/914?