@gfoidl 2018-11-03
In the 32-bit variant this pattern ("setb al / movzx eax, al") doesn't show up, because of the explicit jumps from the comparison. For 32-bit I can just view the dasm in [SharpLab](https://sharplab.io/#v2:C4LghgzgtgPgAgBgARwIwG4CwAoRLUB0ASgK4B2wAllAKYEDCA9lAA6UA2NATgMrcBulAMY0IWbDjgBmFACYk9HAG8cSNShkAjRo3ZIAkhCJgyAcxr6yAIUbkAJhAD6AdUrAAFgAkwQgNYAKSgokRgAzUIgaYAAaJCDgJE4zD1j4pDB+MA4wTU4AGRpk9wBKHABIFWx1JDKUAHYkf38SeOKwiKikAD5GlopijKz2HPzC0w8kGBhe1qTx926Z/v9B7NyaAqKkAFoQ8MjgYuKkAH4kULB2SKQQJGAuEhpxMoBfVXV3tWkkbV0DIxM5ksNnsEFQgWC7QOqWCcxScWCq2G60281K2Aqnxq9SQAEImn1DlDOj1mq0kSMNmMJlMloc4QtSYTiitMmtRltdsTDsVnm9sPygA===) (have only a setup for x64). Here the `sbb + inc`-part is interesting, but I can't judge the cost of this compared to the version with the `? false : true`-hack.