@gfoidl 2018-11-03
When `IsRangeInBounds` is written without `#if BIT64` (for use in "general" libraries) as

```c#
[MethodImpl(MethodImplOptions.AggressiveInlining)]
internal static bool IsRangeInBounds(int offset, int length, int availableLength)
{
    if (IntPtr.Size == 8)
    {
        // ...
        return ((ulong)(uint)offset + (ulong)(uint)length > (ulong)(uint)availableLength) ? false : true;
    }
    else
    {
        return ((uint)offset > (uint)availableLength || (uint)length > (uint)(availableLength - offset)) ? false : true;
    }
}
```

"setb al / movzx eax, al" does not vanish. So if this methods is added, I propose to make it public, in order that other places can benefit from this trick.