@mikedn 2018-11-05
> // We don't use this same mechanism in a 32-bit process due to the overhead of 64-bit arithmetic.

@GrabYourPitchforks I'm curious how big the overhead is. Did you measure it?